packages/forge-runners/pyproject.toml
[project]
name = "forge-runners"
version = "0.1.0"
description = "Diffusion model inference runners for local validation and Kubernetes jobs"
requires-python = ">=3.11,<3.13"
dependencies = [
    "forge-core",
    "forge-eval",
    "diffusers>=0.31",
    "transformers>=4.44",
    "accelerate>=1.0",
    "torch>=2.5,<2.8",
    "pillow>=11.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.scripts]
forge-run = "forge.runners.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]

[tool.uv.sources]
forge-core = { workspace = true }
forge-eval = { workspace = true }
